// src/gen/models/ElementAiGenerated.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI assistant API model
 * This is the API for the AI assistant module
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@obdasystems.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ElementAiGenerated
 */
export interface ElementAiGenerated {
    /**
     * 
     * @type {string}
     * @memberof ElementAiGenerated
     */
    chunkId: string;
    /**
     * 
     * @type {boolean}
     * @memberof ElementAiGenerated
     */
    isNew?: boolean;
}

/**
 * Check if a given object implements the ElementAiGenerated interface.
 */
export function instanceOfElementAiGenerated(value: object): value is ElementAiGenerated {
    if (!('chunkId' in value) || value['chunkId'] === undefined) return false;
    return true;
}

export function ElementAiGeneratedFromJSON(json: any): ElementAiGenerated {
    return ElementAiGeneratedFromJSONTyped(json, false);
}

export function ElementAiGeneratedFromJSONTyped(json: any, ignoreDiscriminator: boolean): ElementAiGenerated {
    if (json == null) {
        return json;
    }
    return {
        
        'chunkId': json['chunkId'],
        'isNew': json['isNew'] == null ? undefined : json['isNew'],
    };
}

export function ElementAiGeneratedToJSON(json: any): ElementAiGenerated {
    return ElementAiGeneratedToJSONTyped(json, false);
}

export function ElementAiGeneratedToJSONTyped(value?: ElementAiGenerated | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'chunkId': value['chunkId'],
        'isNew': value['isNew'],
    };
}