// src/gen/models/RDFGraph.ts
export interface RDFGraph {
    /**
     * 
     * @type {Array<Diagram>}
     * @memberof RDFGraph
     */
    diagrams: Array<Diagram>;
    /**
     * 
     * @type {Array<GrapholscapeEntity>}
     * @memberof RDFGraph
     */
    entities: Array<GrapholscapeEntity>;
    /**
     * 
     * @type {Array<ClassInstanceEntity>}
     * @memberof RDFGraph
     */
    classInstanceEntities?: Array<ClassInstanceEntity>;
    /**
     * 
     * @type {RDFGraphMetadata}
     * @memberof RDFGraph
     */
    metadata: RDFGraphMetadata;
    /**
     * 
     * @type {RDFGraphConfig}
     * @memberof RDFGraph
     */
    config?: RDFGraphConfig;
    /**
     * 
     * @type {number}
     * @memberof RDFGraph
     */
    selectedDiagramId?: number;
    /**
     * 
     * @type {string}
     * @memberof RDFGraph
     */
    modelType: RDFGraphModelTypeEnum;
    /**
     * 
     * @type {Array<Action>}
     * @memberof RDFGraph
     */
    actions?: Array<Action>;
    /**
     * 
     * @type {string}
     * @memberof RDFGraph
     */
    creator?: string;
    /**
     * 
     * @type {Array<SHACLShape>}
     * @memberof RDFGraph
     */
    constraints?: Array<SHACLShape>;
}

/**
* @export
* @enum {string}
*/
export enum RDFGraphModelTypeEnum {
    ONTOLOGY = 'ontology',
    VKG = 'vkg',
    RDF = 'rdf'
}